# swagger/swagger.json
{"swagger":"2.0","info":{"title":"news api sample","description":"goaを用いたnews apiのサンプル","contact":{"name":"shu920921","email":"m153304w@st.u-gakugei.ac.jp","url":"https://github.com/shu920921/goa_news_api/issues"},"license":{"name":"MIT"},"version":""},"host":"localhost:18080","schemes":["http","https"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/api/v1/articles":{"get":{"tags":["articles"],"summary":"articles articles","description":"記事の一覧","operationId":"articles#articles","produces":["application/vnd.goa.error","application/vnd.goa.news.api.article+json; type=collection"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaNewsApiArticleCollection"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http","https"]}},"/api/v1/articles/{id}":{"get":{"tags":["articles"],"summary":"article articles","description":"記事の詳細","operationId":"articles#article","produces":["application/vnd.goa.error","application/vnd.goa.news.api.article+json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaNewsApiArticle"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http","https"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http","https"]}},"/swagger/{filepath}":{"get":{"summary":"Download public/swagger/","operationId":"swagger#/swagger/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http","https"]}}},"definitions":{"GoaNewsApiArticle":{"title":"Mediatype identifier: application/vnd.goa.news.api.article+json; view=default","type":"object","properties":{"clicks":{"type":"integer","example":3317073926577087682,"format":"int64"},"description":{"type":"string","example":"Ratione odio magnam."},"head_line":{"type":"string","example":"Qui est ipsam."},"id":{"type":"integer","example":7190905728174022485,"format":"int64"}},"description":"記事詳細のデータ (default view)","example":{"clicks":3317073926577087682,"description":"Ratione odio magnam.","head_line":"Qui est ipsam.","id":7190905728174022485},"required":["id","head_line","description","clicks"]},"GoaNewsApiArticleCollection":{"title":"Mediatype identifier: application/vnd.goa.news.api.article+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/GoaNewsApiArticle"},"description":"GoaNewsApiArticleCollection is the media type for an array of GoaNewsApiArticle (default view)","example":[{"clicks":3317073926577087682,"description":"Ratione odio magnam.","head_line":"Qui est ipsam.","id":7190905728174022485}]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"NotFound":{"description":"Not Found"}}}